+++ src/components/Modal/Modal.stories.js
@@ -1,3 +1,4 @@
+import Dropdown from '../Dropdown/Dropdown.vue';
 import LobButton from '../Button/Button.vue';
 import RadioButton from '../RadioButton/RadioButton.vue';
 import RadioGroup from '../RadioGroup/RadioGroup.vue';
@@ -30,12 +31,10 @@ const PrimaryTemplate = (args, { argTypes }) => ({
     <Modal
       v-bind="args"
       :visible="isModalVisible"
+      header="Select Tracking Events"
+      closeButtonAriaLabel="Close Tracking Events Modal"
       @close="isModalVisible = false"
     >
-      <template v-slot:header>
-        <h4>Select Tracking Events</h4>
-      </template>
-
       Would you like to export an additional CSV of associated tracking events?
       <RadioGroup>
         <radio-button name="exportCSV" id="yes" value="yes" label="Yes" v-model="radioModel"/>
@@ -52,6 +51,51 @@ const PrimaryTemplate = (args, { argTypes }) => ({
     `
 });
 
+const dropVModel = '';
+const WithDropdownTemplate = (args, { argTypes }) => ({
+  props: Object.keys(argTypes),
+  components: { Modal, LobButton, Dropdown },
+  setup: () => ({ args }),
+  data: () => ({ isModalVisible, dropVModel }),
+  template: `
+    <LewButton @click="isModalVisible = true">
+      Open Modal
+    </LobButton>
+    <Modal
+      v-bind="args"
+      width="500px"
+      :visible="isModalVisible"
+      header="A Modal with a Dropdown"
+      closeButtonAriaLabel="Close modal with dropdown"
+      @close="isModalVisible = false"
+    >
+      <div style="height: 150px;">
+        <div class="mb-5">Select a thing to continue:</div>
+        <Dropdown 
+          id="dropdown1" 
+          label="thing"
+          srOnlyLabel 
+          placeholder="Select a value"
+          :options="['one', 'two']" 
+          v-model="dropVModel"/>
+      </div>
+
+      <template v-slot:footer>
+        <div class="flex self-end">
+          <LobButton 
+          class="ml-2" 
+          :disabled="!dropVModel"
+          @click="isModalVisible=false">OK</LobButton>
+        </div>
+      </template>
+    </Modal>
+    `
+});
+
 export const Primary = PrimaryTemplate.bind({});
 Primary.args = {
+};
+
+export const WithDropdown = WithDropdownTemplate.bind({});
+WithDropdown.args = {
 };