+++ src/components/Modal/Modal.vue
@@ -3,46 +3,47 @@
     <div
       v-show="visible"
       :class="['fixed inset-0 flex items-center justify-center bg-black bg-opacity-60 z-30']"
+      :aria-hidden="!visible"
       @mousedown="closeModal"
+      @keydown.esc="closeModal"
     >
       <div
-        class="bg-white flex flex-col overflow-y-auto shadow rounded-lg p-5 max-h-5/6"
         role="dialog"
-        title="modal"
-        aria-labelledby="modal"
+        aria-modal="true"
+        aria-labelledby="header"
+        aria-describedby="modalDescription"
         :style="{'width': width}"
+        class="relative bg-white flex flex-col overflow-y-auto shadow rounded-lg p-5 max-h-5/6"
         @mousedown.stop
       >
         <header
-          v-if="hasHeader"
-          id="modalTitle"
-          class="flex justify-between border-b border-gray-100 pb-4"
+          v-if="header"
+          id="header"
+          class="pageheading border-b border-gray-100 pb-4"
         >
-          <slot name="header" />
-          <button
-            :class="['rounded-full w-7 h-7 p-1 cursor-pointer hover:bg-white-200',
-                     'focus:outline-none focus:ring-2 focus:ring-primary-100']"
-            aria-label="Close modal"
-            @click="closeModal"
-            @keyup.enter="closeModal"
-          >
-            <Close class="w-5 h-5" />
-          </button>
+          {{ header }}
         </header>
-
         <section
           id="modalDescription"
           class="py-5"
         >
           <slot />
         </section>
-
         <footer
           v-if="hasFooter"
           class="flex border-t border-gray-100 flex-col pt-4"
         >
           <slot name="footer" />
         </footer>
+        <button
+          :class="['absolute top-6 right-4 rounded-full w-7 h-7 p-1 cursor-pointer hover:bg-white-200',
+                   'focus:outline-none focus:ring-2 focus:ring-primary-100']"
+          :aria-label="closeButtonAriaLabel"
+          @click="closeModal"
+          @keyup.enter="closeModal"
+        >
+          <Close class="w-5 h-5" />
+        </button>
       </div>
     </div>
   </transition>
@@ -61,13 +62,18 @@ export default {
     width: {
       type: String,
       default: ''
+    },
+    header: {
+      type: String,
+      required: true
+    },
+    closeButtonAriaLabel: {
+      type: String,
+      required: true
     }
   },
   emits: ['close'],
   computed: {
-    hasHeader () {
-      return Boolean(this.$slots.header);
-    },
     hasFooter () {
       return Boolean(this.$slots.footer);
     }