+++ src/components/Modal/__tests__/Modal.spec.js
@@ -4,25 +4,42 @@ import Modal from '../Modal.vue';
 
 const renderComponent = (options) => render(Modal, { ...options });
 
+const initialProps = { header: 'modal header', closeButtonAriaLabel: 'close this modal' };
+
 describe('Modal', () => {
 
   it('is hidden by default', () => {
-    const { queryByRole } = renderComponent();
-
-    const modal = queryByRole('dialog');
-    expect(modal).not.toBeInTheDocument();
-
-  });
-
-  it('is visible when its visible prop is true', () => {
-    const props = {
-      visible: true
-    };
-
+    const props = initialProps;
     const { queryByRole } = renderComponent({ props });
 
     const modal = queryByRole('dialog');
-    expect(modal).toBeInTheDocument();
+    expect(modal).not.toBeInTheDocument();
+  });
+
+  describe('when visible (visible prop is true)', () => {
+
+    let component; let props;
+    beforeEach(() => {
+      props = {
+        ...initialProps,
+        visible: true
+      };
+      component = renderComponent({ props });
+    });
+
+    it('the modal is visible', () => {
+      const { queryByRole } = component;
+
+      const modal = queryByRole('dialog');
+      expect(modal).toBeVisible();
+    });
+
+    it('displays the header', () => {
+      const { getByText } = component;
+
+      const modalHeader = getByText(props.header);
+      expect(modalHeader).toBeVisible();
+    });
 
   });